+++ packages/pluggableWidgets/datagrid-web/src/components/Table.tsx
@@ -23,8 +23,8 @@ import { ThreeStateCheckBox } from "@mendix/widget-plugin-grid/components/ThreeS
 import { MultiSelectionStatus } from "@mendix/widget-plugin-grid/selection";
 import { SelectionMethod } from "../features/selection";
 import { StickyHeaderTable } from "./StickyHeaderTable";
-import { GridColumn } from "../helpers/GridColumn/GridColumn";
-import { sortColumns } from "../helpers/GridColumn/utils";
+import { GridColumn } from "../helpers/typings/GridColumn";
+import { sortColumns } from "../helpers/utils";
 import { CellComponent } from "../../typings/CellComponent";
 import { Row } from "./Row";
